Change Freeze dialog box to include the highlighted organism (this function does not yet work)
Top of main window now shows the name of the current workspace.

Added the freeze button to the front of the petri dish, attached this to the
freeze item dialog control.

Added the snowflake.png file -- icon for the new freeze button.

Centered the message in the quit dialog.

Changed word "repopulate" to "start new experiment" when user drags new 
population onto the petri dish