Attempted to add new dialog box forcing the user to choose an existing or open a new workspace -- does not work at this moment.
Put in a hack to overcome problem where code Avida-ED crashes if Avida core ends
(for instance if ends on number of updates).  Now the program will pause when 
the user selects that option.

No longer askes the user for a random number seed.  Now they have a choice of
making the run repeatable (random number seed set to 777) or not (seed set 
to 0).

Rearranged the tasks rewarded part of the configuration pane.  They are now
listed as five increasingly hard classes.